% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMPosterior.R
\name{IMPosterior}
\alias{IMPosterior}
\title{IMPosterior}
\usage{
IMPosterior(prior = NULL, posterior = NULL, MME = 0,
  threshold = 0.75, units = NULL, colors, width = NULL,
  height = NULL, elementId = NULL)
}
\arguments{
\item{prior}{draws from the prior distribution. This is an optional parameter but either this or posterior should be provided.}

\item{posterior}{draws from the posterior distribution. This is an optional parameter but either this or prior should be provided.}

\item{MME}{minimum meaninful effect. If not provided MME is set to zero.}

\item{threshold}{if the probability is greater than this threshold, you would feel confortable making a decision}

\item{units}{optional arguement to specify the units of x. For example, dollars or applications.}

\item{colors}{colors for the left, middle, and right areas. The defaults are c("#e41a1c", "#377eb8", "#4daf4a")}

\item{width}{width for shiny}

\item{height}{height for shiny}

\item{elementId}{elementID for shiny}
}
\value{
htmlwidget
}
\description{
IMPosterior
}
\examples{
if(interactive()){
set.seed(9782)
library(IMPosterior)
IMPosterior(prior= rnorm(100000))
 }
}
